Add a min filesize for caching.
git-svn-id: http://svn.scratchpads.eu/svn/fencedine@15 7442ffe0-38b0-4df4-8543-a0eeb3e43e7c

+++ index.php
@@ -1,10 +1,11 @@
 <?php
 
 // ------------------------ SETTINGS -------------------------------------------
-$fencedine_dir  = '/var/www/fencedine.myspecies.info/files';
-$fencedine_path = '/files';
-$fencedine_log  = 'urls';
-$fencedine_age  = 86400;
+$fencedine_dir  = '/var/www/fencedine.myspecies.info/files'; // Directory path
+$fencedine_path = '/files'; // URL path
+$fencedine_log  = 'urls'; // Log file
+$fencedine_age  = 86400; // Max age of cache
+$fencedine_min  = 10; // Minimum size of file to cache
 // -----------------------------------------------------------------------------
 
 /**
@@ -28,8 +29,8 @@ $filename = "$fencedine_dir/$md5_url";
 // Do the redirect.
 header("Location: $fencedine_path/$md5_url");
   
-// If the file exists, redirect to it.
-if(file_exists($filename)){  
+// If the file exists and is bigger than min_bytes, redirect to it.
+if(file_exists($filename) && filesize($filename)>=$fencedine_min){  
   // If the file is older than x seconds, then silently recreate the file
   if(filectime($filename) < time()-$fencedine_age){
     exec('nohup wget --quiet "' . $_GET['url'] . '" -O ' . $filename . ' > /dev/null & echo $!');
@@ -38,6 +39,20 @@ if(file_exists($filename)){
   // The file doesn't exist, we need to download it and wait for it to be
   // downloaded before redirecting.
   exec('wget --quiet "' . $_GET['url'] . '" -O ' . $filename . ' > /dev/null');
+  // Just for logging purposes (the following can be uncommented), I'm saving
+  // a list of URLs accessed
+  file_put_contents($fencedine_log, $_GET['url']."\n", FILE_APPEND);
+}
+
+  
+  // If the file is older than x seconds, then silently recreate the file
+  if(filectime($filename) < time()-$fencedine_age){
+    exec('nohup wget "' . $_GET['url'] . '" -O ' . $filename . ' > /dev/null & echo $!');
+  }
+} else {
+  // The file doesn't exist, we need to download it and wait for it to be
+  // downloaded before redirecting.
+  exec('wget "' . $_GET['url'] . '" -O ' . $filename . ' > /dev/null');
   // Just for logging purposes (the following can be uncommented), I'm saving
   // a list of URLs accessed
   file_put_contents($fencedine_log, $_GET['url']."\n", FILE_APPEND);
